requirements.txt
# Core parsing and NLP
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.4

# Text processing utilities
rapidfuzz==3.9.6
regex==2024.9.11

# Database & JSON handling
sqlite-utils==3.36
pandas==2.2.3

# Optional enhancements
nltk==3.9.1
pyresparser==1.0.6   # optional if using advanced resume parsing
